bug fix refactoring and protocol change
- the protocol has change : the pacth action must tell if the position
or length it send is in byte or char.
the text sent for inserting must be in utf8.
- if the action indicate that the patch are in character the patcher
will calculate the offset in byte otherwise it is passed as is.
Action syntax is now (as a regexp):
"@([bc])([+-])([0-9]+):([^\x01]+)\x01"

(*
[bc] indicate if position and length are in byte or characters
[+-] indicate if we want delete (-) or insert (+)
next group is the position in the source
next group is length if it's a delete operation otherwise it's the text
to be inserted
*)